requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fire==0.7.0
gunicorn==23.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.2
Pygments==2.19.1
PyJWT==2.9.0
PyYAML==6.0.2
rav==0.0.9
rich==13.9.4
sqlparse==0.5.3
termcolor==2.5.0
whitenoise==6.9.0
